moved server implementation into server directory and implemented client for request
